fix most compiler warnings

+++ battleLogic.h
@@ -37,14 +37,14 @@ public:
 	void OnCurrentMonsterDeath();
 
 	// army-wide data
-	size_t lost;
-	size_t armySize;
+	int8_t lost;
+	int8_t armySize;
 	int8_t *lineup;
 
 	int frontDamageTaken;
 	int healing;
 	int cumAoeDamageTaken;
-	float berserkProcs;
+	int berserkProcs;
 
 	// skill data
 	int damage;
